web-project/requirements.txt
certifi==2019.6.16
Django==2.2.4
djangorestframework==3.10.2
mysqlclient==1.3.14
numpy==1.17.0
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
wincertstore==0.2
